+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the YoWASP toolchain extension will be documented in this file.
 
+## 0.1.1 (unreleased)
+
+- Added extraction of bundle entry point via package.json
+- Fixed a race condition in file input/output
+
 ## 0.1.0
 
 - Initial release